\name{SageBionetworksCoex-internal}
\title{Internal SageBionetworksCoex objects}
\description{Internal SageBionetworksCoex objects.}
\details{These are not to be called by the user.}
\alias{computeClusterCoefficient}
\alias{computeModuleCC}
\alias{computeModuleLinks}
\alias{cutree2}
\alias{mergeClusterByPCA}
\alias{moduleDetectLabel}
\alias{moduleHubProfiles}
\alias{openImgDev}
\alias{.computeLinksInNeighbors}
\alias{.getDisplayColorSequence}
\alias{.getNameOfMaxElement}
\alias{.reassignModuleNames}
\alias{.assignModuleColor}
\alias{.getFileExtension}
\alias{.moduleDetectByFixedClusterno}
\keyword{internal}